Add config file

+++ handlers/notes.py
@@ -1,25 +1,29 @@
-from aiogram import types
+from aiogram import types, Bot
 from db import funcs
 
-nsave = '💼 <b>Заметка </b><code>{}</code><b> сохранена!</b>'
-nstop = '💼 <b>Заметка </b><code>{}</code><b> удалена!</b>'
-answer = '💼 <b>Заметки:</b>\n'
+from config import STRING
 
-async def notes_handler(event: types.Message):
+nsave = STRING['nsave']
+nstop = STRING['nstop']
+answer = STRING['notes']
+
+async def notes_handler(event: types.Message, bot: Bot):
     if event.chat.id != event.from_user.id:
         args = event.text.split(' ')
         command = args[0]
-        if len(args) == 2:
-            if command == '/nsave' and event.reply_to_message:
-                funcs.create_note(
-                    event.chat.id, 
-                    args[1], 
-                    event.reply_to_message.text
-                )
-                await event.answer(nsave.format(args[1]))
-            if command == '/nstop':
-                funcs.delete_note(event.chat.id, args[1])
-                await event.answer(nstop.format(args[1]))
+        member = await bot.get_chat_member(event.chat.id, event.from_user.id)
+        if 'administrator' == member.status or 'creator' == member.status:
+            if len(args) == 2:
+                if command == '/nsave' and event.reply_to_message:
+                    funcs.create_note(
+                        event.chat.id, 
+                        args[1], 
+                        event.reply_to_message.text
+                    )
+                    await event.answer(nsave.format(args[1]))
+                if command == '/nstop':
+                    funcs.delete_note(event.chat.id, args[1])
+                    await event.answer(nstop.format(args[1]))
         if len(args) == 1:
             if command == '/notes':
                 msg = answer
